Added input management
Added an input manager, input state, and input mapping classes.

CInputManager hold everthing and is a singleton.

CInputState holds all of the actions for a given input state. The purpose of splitting actions into different states is because it'll be easier to recognize if there's an input conflict when the same button is mapped to separate things.

CInputMapping holds each device input that's mapped to its action. This is where we actually check if we've received any inputs.

Also I've removed the list json files. It's a bit annoying making sure all desired names and paths are added to those files. Instead I just read the images, sprites, and collections folder, make their name be the file name, and then tie it to its file path. This way all I have to do is put something in those folders to be able to call them in the game.