Shadows and alignment
Added some shadow settings to allow 3d sprites to cast and receive shadows, along with some added field for the CSettings class.

Added alignment handling so that when the size of the window is changed, sprites will still be properly aligned with whatever side or corner they've specified. This alignment is separate from offset. Right now offset is not implemented. The position of the 2d sprite is still in the top left corner of the sprite. I will implement this if I feel it's actually needed.