Added script pausing
Added the ability to pause a playing script. This can be done with a Stop() function or the Play() function if the current conflicting animation is the one you're trying to play, and the stop type is pause. Calling the Play() function on an animation that is currently paused and passing in the pause end type will resume that animation.

Changed the conflicting animation organization to lists instead of modified values. The sprite data files will now have a two dimensional array of animations. All animations in the same row will conflict with one another.