Change the vector classes to template classes
Each vector class now has a template class that supports floats and ints.

There's one very weird problem I ran into with the GetWH function in the parse helpers. For some reason the declaration of this function in the header file can't find its definition in the cpp after compiling. This isn't an issue for any of my other parse functions that take in a template class. So I don't know what makes this one different.