Menus, controls, visibility, and deletion.
CMenu class has been added, though not yet functional.

The CControl class has been renamed to iControl and is closer to becoming functional.

Added the CMenuManager singleton class. It is also not yet functional.

Added inherited visibility with parents. If a parent is hidden, its children will also be hidden. If the parent is visible, it's children will be reset with whatever their individual visibility settings are.

Added the ability to mark an object for deletion. This will delete an object from the sprite manager once it hits the Update() function. Also any child whose parent is marked for deletion will also be marked for deletion.